refactor: change "home" to "events" on ui
