+++ app/src/main/java/com/example/eventapp/ui/events/EventsViewModel.java
@@ -1,16 +1,16 @@
-package com.example.eventapp.ui.home;
+package com.example.eventapp.ui.events;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class EventsViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public EventsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("Events");
     }
 
     public LiveData<String> getText() {
