+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -4,7 +4,7 @@
     <item
         android:id="@+id/navigation_home"
         android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
+        android:title="@string/title_events" />
 
     <item
         android:id="@+id/navigation_dashboard"
