+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">EventApp</string>
-    <string name="title_home">Home</string>
+    <string name="title_events">Events</string>
     <string name="title_scanqr">Scan Qr</string>
     <string name="title_profiles">Profiles</string>
 </resources>
